# Add pool for bitwriters